fix betty style

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,13 +1,13 @@
 #include "lists.h"
 
 /**
- * print_listint - prints all the elements of a listin_t int
- * @h - pointer to the first element
+ * print_listint - prints all the elements of a listint_t int
+ * @h: pointer to the first element
  * Return: the number of nodes
  */
 size_t print_listint(const listint_t *h)
 {
-	size_t county = 0;
+	unsigned int county = 0;
 
 	while (h != NULL)
 	{
